+++ rs/benchmarks/src/vacuum.rs
@@ -34,10 +34,13 @@ fn vacuum(c: &mut Criterion) {
     }
 
     group.bench_with_input(BenchmarkId::new("Vacuum", 10000), &10000, |bencher, _| {
+        let rt = tokio::runtime::Runtime::new().unwrap();
         bencher.iter_with_setup(
             || {
                 // Remove everything under base_directory
-                std::fs::remove_dir_all(&base_directory).unwrap();
+                if std::path::Path::new(&base_directory).exists() {
+                    std::fs::remove_dir_all(&base_directory).unwrap();
+                }
 
                 // init the collection
                 Collection::<NoQuantizerL2>::init_new_collection(
@@ -47,32 +50,37 @@ fn vacuum(c: &mut Criterion) {
                 .unwrap();
                 let reader =
                     CollectionReader::new(collection_name.to_string(), base_directory.clone());
-                let collection = reader.read::<NoQuantizerL2>().unwrap();
+                let collection = rt.block_on(reader.read::<NoQuantizerL2>()).unwrap();
                 let mut doc_id = 0;
-                vectors.iter().for_each(|vector| {
-                    collection
-                        .insert_for_users(
-                            &user_ids,
-                            doc_id,
-                            vector,
-                            0,
-                            DocumentAttribute::default(),
-                        )
-                        .unwrap();
-                    doc_id += 1;
+                rt.block_on(async {
+                    for vector in vectors.iter() {
+                        collection
+                            .insert_for_users(
+                                &user_ids,
+                                doc_id,
+                                vector,
+                                0,
+                                DocumentAttribute::default(),
+                            )
+                            .await
+                            .unwrap();
+                        doc_id += 1;
+                    }
                 });
-                let segment_name = collection.flush().unwrap();
-                let pending_segment = collection.init_optimizing(&vec![segment_name]).unwrap();
+                let segment_name = rt.block_on(collection.flush()).unwrap();
+                let pending_segment = rt
+                    .block_on(collection.init_optimizing(&vec![segment_name]))
+                    .unwrap();
                 for doc_id in doc_ids_to_delete.iter() {
-                    collection.remove(user_ids[0], *doc_id, 0).unwrap();
+                    rt.block_on(collection.remove(user_ids[0], *doc_id, 0))
+                        .unwrap();
                 }
                 (collection, pending_segment)
             },
             |collection_and_segment| {
                 let (collection, pending_segment) = collection_and_segment;
                 let optimizer = VacuumOptimizer::<NoQuantizerL2>::new();
-                collection
-                    .run_optimizer(&optimizer, &pending_segment)
+                rt.block_on(collection.run_optimizer(&optimizer, &pending_segment))
                     .unwrap();
             },
         );
